# front/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "新闻资讯发布与推荐系统 API文档",
    "description": "我的新闻资讯系统后台接口文档",
    "contact": {
      "name": "Tong",
      "email": "tong@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/system/like": {
      "put": {
        "tags": [
          "news-like-controller"
        ],
        "operationId": "edit_8",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsLike"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "news-like-controller"
        ],
        "operationId": "add_8",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsLike"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/news/export": {
      "post": {
        "tags": [
          "新闻资讯"
        ],
        "operationId": "export_7",
        "parameters": [
          {
            "name": "news",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/News"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/system/like/export": {
      "post": {
        "tags": [
          "news-like-controller"
        ],
        "operationId": "export_8",
        "parameters": [
          {
            "name": "newsLike",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/NewsLike"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/system/user/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getInfo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/tagRel/{id}": {
      "get": {
        "tags": [
          "news-tag-rel-controller"
        ],
        "operationId": "getInfo_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/tag/{id}": {
      "get": {
        "tags": [
          "news-tag-controller"
        ],
        "operationId": "getInfo_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/sync/{newsId}": {
      "get": {
        "tags": [
          "redis-sync-temp-controller"
        ],
        "operationId": "getInfo_3",
        "parameters": [
          {
            "name": "newsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/statistics/{id}": {
      "get": {
        "tags": [
          "news-statistics-controller"
        ],
        "operationId": "getInfo_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/role/{id}": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "operationId": "getInfo_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/operlog/{id}": {
      "get": {
        "tags": [
          "sys-operation-log-controller"
        ],
        "operationId": "getInfo_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/news/{id}": {
      "get": {
        "tags": [
          "新闻资讯"
        ],
        "operationId": "getInfo_7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/news/list": {
      "get": {
        "tags": [
          "新闻资讯"
        ],
        "summary": "查询新闻列表(管理端)",
        "operationId": "list_7",
        "parameters": [
          {
            "name": "news",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/News"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TableDataInfo"
                }
              }
            }
          }
        }
      }
    },
    "/system/like/{id}": {
      "get": {
        "tags": [
          "news-like-controller"
        ],
        "operationId": "getInfo_8",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
    "/system/like/list": {
      "get": {
        "tags": [
          "news-like-controller"
        ],
        "operationId": "list_8",
        "parameters": [
          {
            "name": "newsLike",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/NewsLike"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TableDataInfo"
                }
              }
            }
          }
        }
      }
    },
    "/system/config/{id}": {
      "get": {
        "tags": [
          "sys-config-controller"
        ],
        "operationId": "getInfo_9",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/comment/{id}": {
      "get": {
        "tags": [
          "评论管理"
        ],
        "operationId": "getInfo_10",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/category/{id}": {
      "get": {
        "tags": [
          "news-category-controller"
        ],
        "operationId": "getInfo_11",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/browse/{id}": {
      "get": {
        "tags": [
          "user-browse-controller"
        ],
        "operationId": "getInfo_12",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/news/{ids}": {
      "delete": {
        "tags": [
          "新闻资讯"
        ],
        "operationId": "remove_7",
        "parameters": [
          {
            "name": "ids",
            "in": "path",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
    "/system/like/{ids}": {
      "delete": {
        "tags": [
          "news-like-controller"
        ],
        "operationId": "remove_8",
        "parameters": [
          {
            "name": "ids",
            "in": "path",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AjaxResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "success": {
            "type": "boolean"
          },
          "warn": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        },
        "additionalProperties": {
          "type": "object"
        }
      },
...
      "News": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "createBy": {
            "type": "string",
            "refType": null
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateBy": {
            "type": "string",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "remark": {
            "type": "string",
            "refType": null
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            },
            "refType": null
          },
          "id": {
            "type": "integer",
            "description": "主键ID",
            "format": "int64",
            "refType": null
          },
          "title": {
            "type": "string",
            "description": "新闻标题",
            "refType": null
          },
          "summary": {
            "type": "string",
            "description": "新闻摘要",
            "refType": null
          },
          "categoryId": {
            "type": "integer",
            "description": "所属分类ID",
            "format": "int64",
            "refType": null
          },
          "coverUrl": {
            "type": "string",
            "refType": null
          },
          "content": {
            "type": "string",
            "refType": null
          },
          "publishStatus": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "shelfStatus": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "publishUserId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "publishTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "isDeleted": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "categoryName": {
            "type": "string",
            "refType": null
          }
        },
        "description": "新闻资讯对象"
      },
      "NewsLike": {
        "type": "object",
        "properties": {
          "createBy": {
            "type": "string",
            "refType": null
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateBy": {
            "type": "string",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "remark": {
            "type": "string",
            "refType": null
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            },
            "refType": null
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "newsId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "likeStatus": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "isDeleted": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        }
      },
...
      "TableDataInfo": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          }
        }
      }
    }
  }
}